Implement the exception handlers and Before save the entities chack whether theu present or not.

+++ services/space-service/src/main/java/com/cromxt/space/service/impl/SpaceServiceImpl.java
@@ -47,6 +47,10 @@ public class SpaceServiceImpl implements SpaceService {
 
     @Override
     public void deleteSpace(String spaceId) {
+        boolean exist = spaceRepository.existsById(spaceId);
+        if(!exist){
+            throw new SpaceUserNotFound("Space user not found with " + spaceId);
+        }
         spaceRepository.deleteById(spaceId);
     }
 }
